refactor: 전역 상태로 변경
